refactor(shell): split desktop window helpers
