+++ shell/src/components/desktop/DesktopDockControls.tsx
@@ -1,0 +1,3 @@
+export { AoedeDockButton } from "./AoedeDockButton";
+export { DockIcon } from "./DockIcon";
+export { TrafficLights } from "./TrafficLights";
